backend/requirements.txt
# ── Core Framework ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

# ── Database ──
sqlalchemy==2.0.36
alembic==1.14.1
aiosqlite==0.20.0

# ── Auth ──
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# ── AI / ML ──
openai-whisper==20240930
openai==1.59.5
langchain==0.3.14
langchain-community==0.3.14
langchain-chroma==0.2.2
chromadb==0.5.23
sentence-transformers==3.3.1

# ── Audio Processing ──
pydub==0.25.1
ffmpeg-python==0.2.0

# ── Export ──
python-docx==1.1.2
reportlab==4.2.5
fpdf2==2.8.2

# ── Task Queue ──
celery==5.4.0
redis==5.2.1

# ── Utilities ──
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
aiofiles==24.1.0

# ── Encryption ──
cryptography==44.0.0
